refactor: Move quota package to sink/tablesink/quota
